refactor(lsp): extract document highlight provider into perl-lsp-document-highlight microcrate
Move the 1,080-line document highlight provider from
crates/perl-lsp/src/features/document_highlight.rs into its own
SRP microcrate perl-lsp-document-highlight. The new crate depends
only on perl-ast (leaf dependency), keeping it lightweight and
independently testable. The original file now re-exports from the
new crate, preserving full backward compatibility.

- New crate: crates/perl-lsp-document-highlight/ (perl-ast dep only)
- 15 tests moved and passing in the new crate
- perl-lsp library (145 tests) continues to pass unchanged
- Added to workspace members, workspace deps, and publish allowlist
