+++ crates/perl-lsp-document-highlight/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "perl-lsp-document-highlight"
+version = "0.11.0"
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "SRP microcrate for Perl LSP document highlight (symbol occurrence highlighting)"
+documentation = "https://docs.rs/perl-lsp-document-highlight"
+keywords = ["perl", "lsp", "document-highlight", "ide", "language-server"]
+categories = ["development-tools", "text-editors"]
+include = [
+  "src/**",
+  "Cargo.toml",
+  "LICENSE*"
+]
+
+[dependencies]
+perl-ast = { workspace = true }
+
+[dev-dependencies]
+perl-parser = { workspace = true }
+
+[lints]
+workspace = true
