+++ crates/perl-lsp/Cargo.toml
@@ -39,6 +39,7 @@ perl-lsp-transport = { workspace = true }
 perl-lsp-code-actions = { workspace = true }
 perl-lsp-completion = { workspace = true }
 perl-lsp-diagnostics = { workspace = true }
+perl-lsp-document-highlight = { workspace = true }
 perl-lsp-inlay-hints = { workspace = true }
 perl-lsp-code-lens = { workspace = true }
 perl-lsp-navigation = { workspace = true, features = ["lsp-compat"] }
